Fix icfg.mem yielding floats (1.0 instead of 1)

+++ ymp/config.py
@@ -572,7 +572,7 @@ class ConfigMgr(object):
 
         div = parse_number("1"+unit)
 
-        return mem / div
+        return int(mem / div)
 
 
 icfg = ConfigMgr()
